+++ tests/test_interface.py
@@ -7,8 +7,10 @@ Note:
 import unittest
 
 from ec import interface
+from ec.modules.classes import HandledException
 
 from support import target_script
+from support.helpers import expect_exception
 
 interface.setBase(target_script)
 
@@ -23,12 +25,17 @@ class TestInterface(unittest.TestCase):
   def test_call(self):
     assert(interface.call('task1 arg1=1') == (1, 2))
     
+  def test_handled_exception(self):
+    assert(expect_exception(lambda: interface.call('task2'), HandledException))
+    
+  def test_unhandled_exception(self):
+    assert(expect_exception(lambda: interface.call('ex'), ZeroDivisionError))
+    
   def test_call_with_input(self):
     from support.helpers import RawInputHook as RIH
     RIH.values(2,3)
     
     assert(interface.call('task1 arg1=1', True) == (1, 2))
-    
     
   def test_resolve(self):
     assert(interface.resolve('task1').Config['name'] == 'task1')
@@ -57,6 +64,16 @@ class TestInterface(unittest.TestCase):
     assert(len(Members.keys()) == 1)
     assert(Members['task1'] is not None)
     assert(Config['desc'] == 'Description for group1')
+  
+  def test_force_config(self): # ToDo: Test ec.interface.force_config. It needs to be tested within a ec script.
+    pass
+    
+  def test_add(self):
+    interface.add(target_script, lambda arg1: arg1, dict(name='added'), [dict(name='arg1', type=int)])
+    
+    assert(interface.call('added arg1=1') == 1)
+    
+    assert(expect_exception(lambda: interface.call('added arg1=a'), HandledException))
     
 if __name__ == '__main__':
   unittest.main()